Renamed modules to shorten lengthy file names.
